move MonthlyStaffingUtil for reusablity
